requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
numpy==2.0.0
pandas==2.2.2
pandasql==0.7.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
SQLAlchemy==2.0.31
tqdm==4.66.4
Werkzeug==3.0.3
zipp==3.19.2
